Fix: Correct data copy size in z_vrfy_espi_read_lpc_request
The z_vrfy_espi_read_lpc_request function in drivers/espi/espi_handlers.c
was incorrectly using sizeof(uint8_t) when copying the result of
z_impl_espi_read_lpc_request (a uint32_t) back to your space.
This would lead to data truncation, with only the least significant
byte of the 32-bit value being copied.

This commit changes sizeof(uint8_t) to sizeof(uint32_t) in the
k_usermode_to_copy call to ensure the full 32-bit value is
correctly copied.

A review of similar z_vrfy_... functions in I2C, SPI, and CAN
handlers did not reveal any other such size mismatches.

+++ drivers/espi/espi_handlers.c
@@ -40,7 +40,7 @@ static inline int z_vrfy_espi_read_lpc_request(const struct device *dev,
 	K_OOPS(K_SYSCALL_DRIVER_ESPI(dev, read_lpc_request));
 
 	ret = z_impl_espi_read_lpc_request(dev, op, &data_copy);
-	K_OOPS(k_usermode_to_copy(data, &data_copy, sizeof(uint8_t)));
+	K_OOPS(k_usermode_to_copy(data, &data_copy, sizeof(uint32_t)));
 
 	return ret;
 }
